Разделил весь код на отдельные js модули
